+++ src/board/board_logic.cpp
@@ -221,3 +221,8 @@ bool CBoardLogic::is_pawn_structure(char white_or_black_pawn, SSquare square1, S
         && ((square4 == NULL_SQUARE) || is_pawn_at(white_or_black_pawn, square4)));
 }
 
+bool CBoardLogic::is_simplified_testcase() {
+    // Both kings missing
+    return ((king_square(WHITE_PLAYER) == NULL_SQUARE) && abs(board.evaluator.nega_score() < SCORE_HALF_KING));
+}
+